+++ lib/all-dependencies.js
@@ -2,8 +2,8 @@
 
 var Dependency    = require('./models/dependency');
 var Descriptor    = require('./models/descriptor');
-var utils         = require('./utils');
-var uniq          = utils.uniq;
+var array         = require('./utils/array');
+var uniq          = array.uniq;
 
 function removeExports(imports) {
   var exportsIndex = imports.indexOf('exports');
@@ -26,15 +26,13 @@ var AllDependencies = {
         descriptor: descriptor,
         dedupedImports: imports
       });
-
-      dependency.addToGraph(importName, graph);
-      dependency.addToImports(importName, imports);
     } else {
       dependency = this._graph[name];
-      dependency.addToGraph(importName, graph);
-      dependency.addToImports(importName, imports);
       dependency.addToDedupedImports(imports);
     }
+
+    dependency.addToGraph(importName, graph);
+    dependency.addToImports(importName, imports);
   },
 
   update: function(descriptor, dependencies) {
@@ -83,9 +81,8 @@ var AllDependencies = {
   for: function(fileOrPackage, parent) {
     var graph = this._graph[fileOrPackage];
 
-
     if (!graph && parent) {
-      return this._graph[parent].imports[fileOrPackage] 
+      return this._graph[parent].imports[fileOrPackage];
     }
 
     return graph;
