+++ lib/models/dependency.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var CoreObject = require('core-object');
-var uniq = require('../utils').uniq;
+var uniq = require('../utils/array').uniq;
 
 module.exports = CoreObject.extend({
   descriptor: {},
@@ -12,7 +12,8 @@ module.exports = CoreObject.extend({
     this.graph[importee] = graph;
   },
   addToDedupedImports: function (imports) {
-    this.dedupedImports = uniq(this.dedupedImports.concat(imports));
+    this.dedupedImports = this.dedupedImports.concat(imports);
+    this.dedupedImports = uniq(this.dedupedImports);
   },
   addToImports: function (importee, imports) {
     this.imports[importee] = imports;
