+++ lib/pre-packager.js
@@ -1,24 +1,16 @@
 'use strict';
 
 var AllDependencies         = require('./all-dependencies');
-var syncForwardDependencies = require('./sync-forward-dependencies');
-var utils                   = require('./utils');
+var syncForwardDependencies = require('./utils/sync-forward-dependencies');
+var array                   = require('./utils/array');
 var CoreObject              = require('core-object');
 var RSVP                    = require('rsvp');
 var walkSync                = require('walk-sync');
 var path                    = require('path');
 var fs                      = require('fs-extra');
-var generateNodeInfo        = require('./nodes/node');
 var mapSeries               = require('promise-map-series');
-var uniq                    = utils.uniq;
-var flatMap                 = utils.flatMap;
-var isEntryFiles            = utils.isEntryFiles;
-var getImportInfo           = require('./get-import-info');
-var zip                     = utils.zip;
-
-function contains(item, arr) {
-  return arr.indexOf(item) > -1;
-}
+var getImportInfo           = require('./utils/get-import-info');
+var zip                     = array.zip;
 
 module.exports = CoreObject.extend({
   init: function(inputTrees, options) {
@@ -104,7 +96,9 @@ module.exports = CoreObject.extend({
   syncForwardEntry: function(entry, relativePaths) {
     var srcDir = this.treeDescriptors[entry].srcDir;
 
-    relativePaths.filter(isEntryFiles(entry)).forEach(function(relativePath) {
+    relativePaths.filter(function(relativePath) {
+      return relativePath.slice(-1) !== '/';
+    }).forEach(function(relativePath) {
       var destination = path.join(this.outputPath, relativePath);
       var source = path.join(srcDir, relativePath);
 
@@ -197,7 +191,6 @@ module.exports = CoreObject.extend({
       if (this.resolutionTypes.indexOf(type) < 0) {
         throw new Error('You do not have a resolver for ' + importInfo.type + ' types.');
       }
-
 
       this.cacheImport(importInfo);
