+++ lib/resolvers/ember-app.js
@@ -2,7 +2,7 @@
 
 var path                    = require('path');
 var AllDependencies         = require('../all-dependencies');
-var syncForwardDependencies = require('../sync-forward-dependencies');
+var syncForwardDependencies = require('../utils/sync-forward-dependencies');
 var fs                      = require('fs-extra');
 
 module.exports = {
