+++ lib/resolvers/npm.js
@@ -1,13 +1,14 @@
 'use strict';
 
-var utils       = require('../utils');
-var arraysEqual = utils.arraysEqual;
+var array       = require('../utils/array');
+var arraysEqual = array.arraysEqual;
 var RSVP        = require('rsvp');
 var browserify  = require('browserify');
 var path        = require('path');
 var fs          = require('fs-extra');
 var mapSeries   = require('promise-map-series');
 var hasForDep   = require('hash-for-dep');
+var getImportInfo = require('../utils/get-import-info');
 
 function generateStub(moduleName) {
   return 'define("npm:' + moduleName + '", function() {' +
@@ -17,7 +18,7 @@ function generateStub(moduleName) {
 
 function hashDeps(imports, fullPath) {
   return imports.map(function(imprt) {
-    return utils.getImportInfo(imprt).pkgName;
+    return getImportInfo(imprt).pkgName;
   }).reduce(function(a, b) {
     if (a.indexOf(b) < 0) {
       a.push(b);
