+++ lib/utils/get-import-info.js
@@ -1,5 +1,5 @@
 'use strict';
-var Import = require('./models/import');
+var Import = require('../models/import');
 
 function getImportInfo(descriptors, packageName, importee, importer) {
   var descriptor = descriptors[packageName];
