+++ lib/utils/sync-forward-dependencies.js
